Drill instructions improved
Changed: Removed the random sleep from initRifleRange meaning that the drill instructions function is able to run on the briefing screen before the mission is started. This does open the possiblity that duplicated range IDs will not be detected but it is very unlikely.
Changed: Small improvements made to the content of some drill instructions.
Changed: RR_HEADSET_CLASSES only needs to be initialised on cliants so has been switched into the if !dedicated block of initCommon function.